NOT WORKING finished implementing prototype; debugging

+++ include/bliss.h
@@ -260,12 +260,12 @@ void bl_initialize_song(struct bl_song * const song);
  * DOC TODO
  *
  */
-float bl_mean(int16_t *sample_array, int nSamples);
+float bl_mean(double *sample_array, int nSamples);
 
 /**
  *
  * DOC TODO
  *
  */
-float bl_variance(int16_t *sample_array, int nSamples);
+float bl_variance(double *sample_array, int nSamples);
 #endif  // BL_BLISS_H_
